apps/faq/serializers.py
from rest_framework import serializers

from faq.models import Faq


class FaqSerializer(serializers.ModelSerializer):
    class Meta:
        model = Faq
        fields = ('id', 'question', 'answer', 'created_at')
