requirements.txt
requests
pyyaml
Django==2.2.9
Pillow==7.0.0
psycopg2-binary==2.8.4
mysqlclient==1.4.6
grab==0.6.41
six==1.14.0
textract==1.6.3
djangorestframework==3.11.0
dj-database-url==0.5.0
coreapi==2.3.3
markdown==3.1.1
django-filter==2.2.0
sorl-thumbnail==12.5.0
django-modeltranslation==0.14.2
celery[redis]==4.4.0
django_celery_results==1.0.4
django_celery_beat==1.4.0
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.3.0
django-cors-headers==3.2.1
ipython==7.9.0
django_extensions==2.2.6
fcm-django==0.3.4
django_tinymce==2.8.0
